requirements.txt
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytz>=2023.3
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
requests>=2.31.0
gtts>=2.4.0
SpeechRecognition>=3.10.0
pytest>=7.0.0
pytest-cov>=4.0.0